requirements.txt
requests
pyyaml
responses